Centralize graph passes into a single file
I've unified the 3 different implementations of cleanup_graph we had, which should keep the same behavior as before. We can make all the uses of it to have the aggressive flag on, but I'm keeping it BC for now to avoid changing too many things at once
